--- src/Demo/CartBundle/Entity/Cart/OrderCartInterface.php
@@ -1,15 +1,0 @@
-<?php
-
-namespace Demo\CartBundle\Entity\Cart;
-
-/**
- * Interface OrderCartInterface
- * @package Demo\CartBundle\Entity
- *
- * Defines a cart that can do checkout and create order from its order items
- *
- */
-interface OrderCartInterface extends CartInterface, CheckoutInterface
-{
-
-}
